+++ application/src/main/java/br/com/project/rest/cashback/service/IDiscoService.java
@@ -14,7 +14,7 @@ public interface IDiscoService {
 
     Page<Disco> findDiscos(Long generoId, Integer page, Integer linesPerPage, String orderBy, String direction);
 
-    Disco findDisco(Integer id);
+    Disco findDisco(Long id);
 
     List<Disco> findListaDiscosPorIds(List<Long> discosIds);
